// Web/src/generated/model/deploymentPlanEdge.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * ReleaseHub 前後端共用的 HTTP API 契約。
 * OpenAPI spec version: 1.0.0
 */
import type { DeploymentPlanEdgeCondition } from './deploymentPlanEdgeCondition';

export interface DeploymentPlanEdge {
  /**
     * @minLength 1
     * @maxLength 255
     */
  from: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  to: string;
  condition: DeploymentPlanEdgeCondition;
}